requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
absl-py=2.1.0=pyhd8ed1ab_0
aom=3.6.0=h6a678d5_0
blas=1.0=mkl
blosc=1.21.3=h6a678d5_0
bottleneck=1.3.7=py311hf4808d0_0
brotli=1.0.9=h5eee18b_8
brotli-bin=1.0.9=h5eee18b_8
brotli-python=1.0.9=py311h6a678d5_8
brunsli=0.1=h2531618_0
bzip2=1.0.8=h5eee18b_6
c-ares=1.19.1=h5eee18b_0
ca-certificates=2024.7.4=hbcca054_0
cairo=1.16.0=hb05425b_5
certifi=2024.7.4=py311h06a4308_0
cfitsio=3.470=h5893167_7
charls=2.2.0=h2531618_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
contourpy=1.2.0=py311hdb19cb5_0
cuda-cudart=12.4.127=0
cuda-cupti=12.4.127=0
cuda-libraries=12.4.0=0
cuda-nvrtc=12.4.127=0
cuda-nvtx=12.4.127=0
cuda-opencl=12.6.37=0
cuda-runtime=12.4.0=0
cuda-version=12.6=3
cycler=0.11.0=pyhd3eb1b0_0
cyrus-sasl=2.1.28=h52b45da_1
dav1d=1.2.1=h5eee18b_0
dbus=1.13.18=hb2f20db_0
eigen=3.4.0=hdb19cb5_0
expat=2.6.2=h6a678d5_0
ffmpeg=4.2.2=h20bf706_0
filelock=3.13.1=py311h06a4308_0
fontconfig=2.14.1=h55d465d_3
fonttools=4.51.0=py311h5eee18b_0
freetype=2.12.1=h4a9f257_0
giflib=5.2.1=h5eee18b_3
glib=2.78.4=h6a678d5_0
glib-tools=2.78.4=h6a678d5_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py311hc9b5ff0_0
gnutls=3.6.15=he1e5248_0
graphite2=1.3.14=h295c915_1
grpcio=1.62.2=py311h6a678d5_0
gst-plugins-base=1.14.1=h6a678d5_1
gstreamer=1.14.1=h5eee18b_1
harfbuzz=4.3.0=hf52aaf7_2
hdf5=1.12.1=h2b7332f_3
icu=73.1=h6a678d5_0
idna=3.7=py311h06a4308_0
imagecodecs=2023.1.23=py311h8105a5c_0
imageio=2.33.1=py311h06a4308_0
importlib-metadata=8.3.0=pyha770c72_0
intel-openmp=2023.1.0=hdb19cb5_46306
jinja2=3.1.4=py311h06a4308_0
joblib=1.4.2=py311h06a4308_0
jpeg=9e=h5eee18b_3
jxrlib=1.1=h7b6447c_2
kiwisolver=1.4.4=py311h6a678d5_0
krb5=1.20.1=h143b758_1
lame=3.100=h7b6447c_0
lazy_loader=0.4=py311h06a4308_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libabseil=20240116.2=cxx17_h6a678d5_0
libaec=1.0.4=he6710b0_1
libavif=0.11.1=h5eee18b_0
libbrotlicommon=1.0.9=h5eee18b_8
libbrotlidec=1.0.9=h5eee18b_8
libbrotlienc=1.0.9=h5eee18b_8
libclang=14.0.6=default_hc6dbbc7_1
libclang13=14.0.6=default_he11475f_1
libcublas=12.4.2.65=0
libcufft=11.2.0.44=0
libcufile=1.11.0.15=0
libcups=2.4.2=h2d74bed_1
libcurand=10.3.7.37=0
libcurl=8.7.1=h251f7ec_0
libcusolver=11.6.0.99=0
libcusparse=12.3.0.142=0
libdeflate=1.17=h5eee18b_1
libedit=3.1.20230828=h5eee18b_0
libev=4.33=h7f8727e_1
libffi=3.4.4=h6a678d5_1
libgcc-ng=14.1.0=h77fa898_0
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libglib=2.78.4=hdc74915_0
libgomp=14.1.0=h77fa898_0
libgrpc=1.62.2=h2d74bed_0
libiconv=1.16=h5eee18b_3
libidn2=2.3.4=h5eee18b_0
libjpeg-turbo=2.0.0=h9bf148f_0
libllvm14=14.0.6=hdb19cb5_3
libnghttp2=1.57.0=h2d74bed_0
libnpp=12.2.5.2=0
libnvfatbin=12.6.20=0
libnvjitlink=12.4.99=0
libnvjpeg=12.3.1.89=0
libopus=1.3.1=h7b6447c_0
libpng=1.6.39=h5eee18b_0
libpq=12.17=hdbd6064_0
libprotobuf=4.25.3=he621ea3_0
libssh2=1.11.0=h251f7ec_0
libstdcxx-ng=14.1.0=hc0a3c3a_0
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libvpx=1.7.0=h439df22_0
libwebp-base=1.3.2=h5eee18b_0
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=h097e994_2
libxml2=2.13.1=hfdd30dd_2
libzopfli=1.0.3=he6710b0_0
llvm-openmp=14.0.6=h9e868ea_0
lz4-c=1.9.4=h6a678d5_1
markdown=3.6=pyhd8ed1ab_0
markupsafe=2.1.3=py311h5eee18b_0
matplotlib=3.8.4=py311h06a4308_0
matplotlib-base=3.8.4=py311ha02d727_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py311h5eee18b_1
mkl_fft=1.3.8=py311h5eee18b_0
mkl_random=1.2.4=py311hdb19cb5_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py311h06a4308_0
mysql=5.7.24=h721c034_2
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=3.3=py311h06a4308_0
numexpr=2.8.7=py311h65dcdc2_0
numpy=1.26.4=py311h08b1b3b_0
numpy-base=1.26.4=py311hf175353_0
opencv=4.10.0=py311hab19f7d_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.5.2=he7f1fd0_0
openssl=3.3.1=h4bc722e_2
packaging=24.1=py311h06a4308_0
pandas=2.2.2=py311ha02d727_0
pcre2=10.42=hebb0a14_1
pillow=10.4.0=py311h5eee18b_0
pip=24.2=py311h06a4308_0
pixman=0.40.0=h7f8727e_1
ply=3.11=py311h06a4308_0
protobuf=4.25.3=py311h7b78aeb_0
pybind11-abi=4=hd3eb1b0_1
pyparsing=3.0.9=py311h06a4308_0
pyqt=5.15.10=py311h6a678d5_0
pyqt5-sip=12.13.0=py311h5eee18b_0
pysocks=1.7.1=py311h06a4308_0
python=3.11.9=h955ad1f_0
python-dateutil=2.9.0post0=py311h06a4308_2
python-tzdata=2023.3=pyhd3eb1b0_0
python_abi=3.11=2_cp311
pytorch=2.4.0=py3.11_cuda12.4_cudnn9.1.0_0
pytorch-cuda=12.4=hc786d27_6
pytorch-mutex=1.0=cuda
pytz=2024.1=py311h06a4308_0
pyyaml=6.0.1=py311h5eee18b_0
qt-main=5.15.2=h53bd1ea_10
re2=2022.04.01=h27087fc_0
readline=8.2=h5eee18b_0
requests=2.32.3=py311h06a4308_0
scikit-image=0.23.2=py311ha02d727_0
scikit-learn=1.5.1=py311ha02d727_0
scipy=1.13.1=py311h08b1b3b_0
seaborn=0.13.2=py311h06a4308_0
setuptools=72.1.0=py311h06a4308_0
sip=6.7.12=py311h6a678d5_0
six=1.16.0=pyhd3eb1b0_1
snappy=1.2.1=h6a678d5_0
sqlite=3.45.3=h5eee18b_0
sympy=1.12=py311h06a4308_0
tbb=2021.8.0=hdb19cb5_0
tensorboard=2.17.1=pyhd8ed1ab_0
tensorboard-data-server=0.7.0=py311h63ff55d_1
threadpoolctl=3.5.0=py311h92b7b1e_0
tifffile=2023.4.12=py311h06a4308_0
tk=8.6.14=h39e8969_0
torchaudio=2.4.0=py311_cu124
torchtriton=3.0.0=py311
torchvision=0.19.0=py311_cu124
tornado=6.4.1=py311h5eee18b_0
typing_extensions=4.11.0=py311h06a4308_0
tzdata=2024a=h04d1e81_0
unicodedata2=15.1.0=py311h5eee18b_0
urllib3=2.2.2=py311h06a4308_0
werkzeug=3.0.3=pyhd8ed1ab_0
wheel=0.43.0=py311h06a4308_0
x264=1!157.20191217=h7b6447c_0
xz=5.4.6=h5eee18b_1
yaml=0.2.5=h7b6447c_0
zfp=1.0.0=h6a678d5_0
zipp=3.20.0=pyhd8ed1ab_0
zlib=1.2.13=h5eee18b_1
zstd=1.5.5=hc292b87_2
